Refactor(exporter): decouple TraceValidator from HTTP framework
